Random Themes for SquirrelMail by Jorey Bump
These themes generate random colors whenever a page is refreshed or a new one is
loaded.

** NOTE:  Question Reality wasn't added since it was 100% random and would
often produce unreadable results.  From the readme, it was listed as "For
demonstration purposes only!"  I hope Jorey doesn't mind. **

Spice of Life: This theme creates contrasting colors for easier readability. I
have also provided 2 variations, one that always provides a dark background and
one that always provides a light background (my personal favorite).

Greenhouse Effect, Kind of Blue, In the Pink, and Shades of Grey: As the titles
suggest, these themes favor a specific color.

Monostochastic: This creates a contrasting 2 color theme. There is only one
color for text, one for the background. Very spartan.

NOTES:

These themes will not interfere with the message highlighting feature provided
by SquirrelMail, but sometimes the results may be unpleasant.

I haven't tried it, but I assume that any colors specified in an external css
file will override the ones provided by any theme.

On a slightly technical note, this code will actually cause 2 themes to be
generated, one for each frame.

The SquirrelMail developers have not formally declared that certain elements of
the color array be used for text vs. background (although that would be nice!),
so I had to make educated guesses during development. I've noticed that
SquirrelMail borrows a background color from the right frame to use as a text
color in the left, to display parent folder names. This sometimes causes
the folder names to be barely legible. Since my code loads a different theme in
each frame, this situation cannot occur. It's something to keep in mind when
you're designing themes, however. Here's a fix, if you find this to be as
annoying as I do:

Change line 172 (in 1.2.0-rc2) of src/left_main.php from:

 $line .= "<FONT COLOR=\"$color[10]\">";

to:

 $line .= "<b><FONT COLOR=\"$color[6]\">";

Hopefully, this will be changed in future versions, and the color array will be
standardized for text and background colors.


git-svn-id: https://svn.code.sf.net/p/squirrelmail/code/trunk@1874 7612ce4b-ef26-0410-bec9-ea0150e637f0