Correct on the fly base64 decode algoritm in case the returned data does not have "proper" lines with the same length. With an old uw server I discovered that lines were split in the middle by \n. This fix will correct it.
The rfc about base64 decoding does not mention anything about line length.
rfc2045 does mention that lines should not be longer then 78 characters but
does not say all the lines (excluding the last of course) must be of the same
size.
There for i could not ignore it and fixed it.


git-svn-id: https://svn.code.sf.net/p/squirrelmail/code/trunk@5415 7612ce4b-ef26-0410-bec9-ea0150e637f0