Somehow (probably during rg=off changes) we did not cache the mailbox-tree which resulted in unneeded imap requests for information we should have cached in the session. With this fix we cache the mailbox-tree again and we are able to do a force refresh.
Forced calls should be done in case of
rename/delete/add/subscribe/unsubscrive  (folders.php) and
in case we have multiple imap connections and create or subscribe to a
mailbox on another client and want to sync the changes with the SM cache.
For the last reason I moved the Refresh Folder List to below the
mailbox-tree to force a refresh and added a new link on top of the
mailbox-tree to do a unseen/total refresh only.


git-svn-id: https://svn.code.sf.net/p/squirrelmail/code/branches/SM-1_4-STABLE/squirrelmail@5608 7612ce4b-ef26-0410-bec9-ea0150e637f0