filters.php wasn't working with UID's so I rewrote this speed killer plugin and hope it's now working again and at least is a bit faster.
It need to be tested very well because during my tests I got 3000 messages
moved to the trash folder because ip 127.0.0.1 was entitled as spam.
Hopefully it doesn't happen again. Be carefull with the var YourHop because
default it's ' ' which will match the last added Received header and quits
looking for the rest of the headers. If the last added header (the first
when you look top down) contains 127.0.0.1 then the message might be deleted
if you enabled all the spam blacklists.

People who actually use this plugin should take a look.

The next time I touch this plugin again strip all the functions and echo a
link to spamassasin, procmail and sieve server side filters.


git-svn-id: https://svn.code.sf.net/p/squirrelmail/code/trunk/squirrelmail@5937 7612ce4b-ef26-0410-bec9-ea0150e637f0