PARTIAL integration of delete_move_next. I've got the toolbar changed around, and delete/move/next functions mostly integrated (commented out redundant hooks in the plugin).
The processing of delete/move is not quite right ... the plugin is
doing some odd things to remove the element from the msgs/msort arrays.
Marc? Thoughts on how/where best to finish delete processing?
Right now it's still borrowing the code from the plugin to finish the
delete.

Comments on the layout? (why I committed as is).

If you want to play with the not quite finished delete action,
uncomment the lines in read_body.php that handle deleted/moved
messages (around lines 785 - 809), and then in
delete_move_next/setup.php, comment the first two 'action' hooks.

I'll finish up the options work and the delete manipulation tomorrow.
Wanted feedback on the link placement, etc. ASAP.

This was one thing we wanted in before frameless SM... so once this goes in
(pending any other urgent things), we'll cut the dev release.

My apologies for not getting to this sooner. :-/


git-svn-id: https://svn.code.sf.net/p/squirrelmail/code/trunk/squirrelmail@6439 7612ce4b-ef26-0410-bec9-ea0150e637f0