Tomas Kuliavas wrote:
>Hi,
>
>You have posted script that converts SquirrelMail data files from files to
>sql on SquirrelMail users mailing list.
>
>http://article.gmane.org/gmane.mail.squirrelmail.user/6706
>
>Can you clarify script copyrights? I would like to include script in
>SquirrelMail package, but SquirrelMail can use only GPL licensed software.
>
>
>
I spoke with Tal ( the co-author ) and we are both quite comfortable
releasing this under the gpl.

We had originally offered to include it in the distro but no one took us
up on the offer.  We haven't used it in 2-3 years so I don't know if it
works with current versions.  I also don't believe either of us have any
interest in maintaining it but we welcome the projects use of it.

Michael


git-svn-id: https://svn.code.sf.net/p/squirrelmail/code/trunk@10109 7612ce4b-ef26-0410-bec9-ea0150e637f0